+++ context/CONTEXT_2026-02-14_10-32-37.md
@@ -1,0 +1,11 @@
+# Context
+
+**File Chosen:** `js/ui/profileModalController.js`
+**Size:** 8159 lines (Grandfathered)
+**Reason:** Largest grandfathered file not decomposed recently. Contains mixed concerns including Admin management and Moderation settings which are distinct from the main profile modal logic.
+
+**Extraction Plan:**
+1.  **Admin Logic**: Extract methods related to Admin/Moderator management (add/remove moderators, whitelist, blacklist, notifications) into `js/ui/profileModal/ProfileAdminController.js`.
+2.  **Moderation Logic**: Extract methods related to Moderation settings (blur threshold, autoplay, mute/hide logic) into `js/ui/profileModal/ProfileModerationController.js`.
+
+**Goal:** Reduce `js/ui/profileModalController.js` by at least 200 lines (likely much more).
