+++ docs/agents/task-logs/daily/2026-02-14_10-57-56_decompose-agent_completed.md
@@ -1,0 +1,21 @@
+# Decompose Agent Completion Log
+
+**Date:** 2026-02-14
+**Agent:** decompose-agent
+**Target:** js/ui/profileModalController.js
+
+## Actions Taken
+1. Identified `js/ui/profileModalController.js` as the largest grandfathered file (8159 lines).
+2. Extracted `js/ui/profileModal/ProfileAdminController.js` (Admin logic).
+3. Extracted `js/ui/profileModal/ProfileModerationController.js` (Moderation logic).
+4. Extracted `js/ui/profileModal/ProfileAdminRenderer.js` (Admin UI rendering logic) to satisfy file size limits.
+5. Refactored `js/ui/profileModalController.js` to delegate to these new controllers.
+6. Implemented backward-compatibility wrappers and property aliases in `ProfileModalController.js` to ensure existing tests pass.
+
+## Verification
+- **Unit Tests:** `npm run test:unit tests/profile-modal-controller.test.mjs` passed (18/18 tests).
+- **Linting:** `npm run lint` passed. File sizes are compliant.
+- **Reduction:** `js/ui/profileModalController.js` reduced from ~8159 to ~6158 lines (-2001 lines).
+
+## Status
+Success.
