+++ todo/TODO_2026-02-14_10-32-37.md
@@ -1,0 +1,18 @@
+# Todo
+
+- [ ] Extract Admin Logic to `js/ui/profileModal/ProfileAdminController.js`
+  - [ ] Move methods: `populateAdminLists`, `renderAdminList`, `normalizeAdminListEntries`, `handleAddModerator`, `handleRemoveModerator`, `handleAdminListMutation`, `runAdminMutation`, `refreshAdminPaneState`, `setAdminLoading`, `clearAdminLists`, `ensureAdminActor`, `storeAdminEmptyMessages`, `normalizeNpubValue`, `describeAdminError`, `notifyAdminError`, `describeNotificationError`, `sendAdminListNotification`.
+  - [ ] Move DOM references: `moderatorSection`, `moderatorEmpty`, `adminModeratorList`, `addModeratorButton`, `moderatorInput`, `adminModeratorsRefreshBtn`, `whitelistSection`, `whitelistEmpty`, `whitelistList`, `addWhitelistButton`, `whitelistInput`, `adminWhitelistRefreshBtn`, `blacklistSection`, `blacklistEmpty`, `blacklistList`, `addBlacklistButton`, `blacklistInput`, `adminBlacklistRefreshBtn`.
+  - [ ] Update imports in `ProfileAdminController.js`.
+  - [ ] Update `ProfileModalController.js` to use `ProfileAdminController`.
+
+- [ ] Extract Moderation Logic to `js/ui/profileModal/ProfileModerationController.js`
+  - [ ] Move methods: `refreshModerationSettingsUi`, `getModerationSettingsService`, `getModerationSettingsDefaults`, `normalizeModerationSettings`, `readModerationInputs`, `applyModerationSettingsControlState`, `areTrustedHideControlsEnabled`, `updateTrustedHideControlsVisibility`, `updateModerationSettingsStatus`, `updateTrustedMuteHideHelperCopy`, `getModerationOverrideEntries`, `normalizeModerationOverrideEntries`, `formatModerationOverrideTimestamp`, `handleModerationOverrideReset`, `refreshModerationOverridesUi`, `getModerationTrustStats`, `updateModerationTrustStats`, `handleModerationSettingsSave`, `handleModerationSettingsReset`.
+  - [ ] Move DOM references: `moderationSettingsCard`, `moderationBlurInput`, `moderationAutoplayInput`, `moderationMuteHideInput`, `moderationSpamHideInput`, `moderationSaveButton`, `moderationResetButton`, `moderationStatusText`, `moderationOverridesList`, `moderationOverridesEmpty`, `moderationTrustedContactsCount`, `moderationTrustedMuteCount`, `moderationTrustedReportCount`, `moderationSeedOnlyIndicator`, `moderationHideControlsGroup`, `moderationHideControlElements`.
+  - [ ] Update imports in `ProfileModerationController.js`.
+  - [ ] Update `ProfileModalController.js` to use `ProfileModerationController`.
+
+- [ ] Verify
+  - [ ] Run `npm run lint`.
+  - [ ] Run `npm run test:unit`.
+  - [ ] Manual check.
